FEAT: more logic

+++ src/vendor_app/app/src/main/kotlin/cz/quanti/android/vendor_app/repository/deposit/dto/AssistanceBeneficiary.kt
@@ -2,7 +2,7 @@ package cz.quanti.android.vendor_app.repository.deposit.dto
 
 class AssistanceBeneficiary(
     val id: Long,
-    val assistanceId: Long,
-    val beneficiaryId: Long,
+    val assistanceId: Int,
+    val beneficiaryId: Int,
     val smartcardSN: String
 )
